Rework readWords and readStrings to make them way faster.
Avoid the loop with a function that keeps track of an accumulator, and
especially avoid using ++ for building a list.  That keeps a giant pile
of thunks around that consumes tons of memory and tons of CPU time.
This way is a little less clear (I think?) but so much faster.  It
actually completes.