Don't use sinkList in unrpm.
sinkList pulls the entire contents of the conduit into memory.  For a
large RPM, that means we could have a whole lot of uncompressed data in
memory.  So instead, use the conduit version of mapM_ to write the files
out to disk one at a time.

In general, it's probably not a great idea to use sinkList anywhere.