Add test for parseSection
parseSection is a composite of other parsing functions and a map.
The rest of the functions have been tested with both positive and
negative scenarios and IMO we don't need a negative scenario here.

The test data also includes some tags where ty is different than
what we expect (probably older RPM format) and they are filtered
from the result. This covers pretty much all cases.