Rename to codec-rpm.
There's already an rpm module, though it has not been updated in quite
some time.  I've also knocked our version number back down to 0.1.  It
was only at 1 to avoid conflicts with that other module while this was
a subtree of bdcs.  That's not going to be the case any longer, and we
never did an actual tagged release of this module anyway.  So it
shouldn't interfere with anything.

I'll be renaming the source repo on github to match, which I am hoping
means it gets checked out into a renamed directory.