Install hlint and hpc-coveralls at the same time as everything else.
This turns them into build dependencies for the test-suite target.
Elsewhere, we are having a problem where hlint gets installed first,
populating the sandbox with versions of its dependencies that satisfy
hlint but may conflict with the library target.  Then when the library
is built, its dependencies may not be satisfiable given the earlier
choices made when installing hlint.

I have also removed happy as a requirement - nothing appears to use it.